Streamline SslStream state validation
